rename ast to dast, resolves #624
